Added mysql wallet migration test and integrated sqlx migration into libvdrtools
